little fixes and remote admin keys

+++ pkg/install.sh
@@ -2,11 +2,12 @@
 set -eu
 
 STOW=$( which stow )
-PKGDIR=$THETAPI_HOME/pkg
+STOW_DIR=$THETAPI_HOME/pkg
 REBOOT_FLAG=$THETAPI_HOME/.rebootreq
+HUGO_URL="https://github.com/gohugoio/hugo/releases/download/v0.54.0/hugo_0.54.0_Linux-ARM.deb"
+DEV_PUBKEY='ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAABAQDXFbPpSVUYrNh1w5CyOjbFvYFjOWgr6lTIf2beKKLzVD/gbWHlp2gtZ9//zgxzUJ2Ml1tZ7vwOSR4vhMqRJ8eNjl6pp2e1jsxUE4ipHBsj2S+VWIiDJ5JYZ4SzNuL4fduASUmeHB+K7Lxe5zw3Ri8+Z0C9XjLwPqri8rR9sirBuZiobINTwu0IJMFrZmCloZ8r1gg2IgulRfT1C0f+P9coYjDkuRa4W1LdRmAsmKOTNG14YEWCQjRL8q4qtWF1hQ1KMBktrpEYh2uhZiKcPAFDlJXxrIEYtmQ8rGqL17a4Z50NhryW/plKLS/mDUHsW5XNPgvr8eILWid2AkT80pfB thetapi@thetanil.com'
 
 echo "## $1 installing..."
-echo "## TODO: ADD RPI CHECKS"
 echo "## TODO: BACKUP STOWED FILES"
 
 case $1 in
@@ -14,11 +15,14 @@ case $1 in
     # TODO: /etc/defaults/keyboard remap caps:escape
     ;;
   "sshkeys")
-    if [ -d $HOME/.ssh ] && [ ! -f $HOME/.ssh/id_rsa ] && [ ! -f $HOME/.ssh/id_rsa.pub ]; then
+    if [ -d $HOME/.ssh ] && \
+       [ ! -f $HOME/.ssh/id_rsa ] && \
+       [ ! -f $HOME/.ssh/id_rsa.pub ]; then
       ssh-keygen -trsa -N "" -f $HOME/.ssh/id_rsa
     else
       echo "could not generate ssh keys (already exist or ~/.ssh missing)"
     fi
+    echo $DEV_PUBKEY >> $HOME/.ssh/authorized_keys2
     ;;
   "git")
     # TODO: make vars
@@ -28,17 +32,19 @@ case $1 in
     ;;
   "hugo")
     # TODO: AMD for deb, ARM for Pi
-    wget -O $HOME/hugo.deb https://github.com/gohugoio/hugo/releases/download/v0.54.0/hugo_0.54.0_Linux-ARM.deb
-    sudo dpkg -i $HOME/hugo.deb
-    rm $HOME/hugo.deb
-    hugo version
-    cd $THETAPI_HOME/ui
-    hugo
+    if [ -f /etc/rpi-issue ]; then
+      wget -O $HOME/hugo.deb $HUGO_URL
+      sudo dpkg -i $HOME/hugo.deb
+      rm $HOME/hugo.deb
+      hugo version
+      cd $THETAPI_HOME/ui
+      hugo
+    fi
     ;;
   "hosts")
     #TODO: Download latest and schedule update job
     sudo rm /etc/hosts
-    sudo $STOW -d $PKGDIR -t /etc hosts
+    sudo $STOW -t /etc hosts
     ;;
   "disable_useless")
     sudo systemctl disable avahi-daemon
@@ -53,74 +59,91 @@ case $1 in
     sudo systemctl stop bluetooth
 
     if [ ! -f /etc/modprobe.d/blacklist-bluetooth.conf ]; then
-      echo '#bluetooth\nblacklist btbcm\nblacklist hci_uart\n' | sudo tee /etc/modprobe.d/blacklist-bluetooth.conf
+      CONF = '#bluetooth\nblacklist btbcm\nblacklist hci_uart\n'
+      echo $CONF | sudo tee /etc/modprobe.d/blacklist-bluetooth.conf
       touch "$REBOOT_FLAG"
     fi
     ;;
   "vim")
-    $STOW -d $PKGDIR -t $HOME vim
+    $STOW -t $HOME $1
     if [ ! -d "${HOME}/.vim/bundle/Vundle.vim" ]; then
       git clone https://github.com/VundleVim/Vundle.vim.git ~/.vim/bundle/Vundle.vim
     fi
     vim +PluginInstall +qall
     ;;
   "dnsmasq")
-    sudo apt install -y dnsmasq
-    sudo $STOW -d $PKGDIR -t /etc dnsmasq
-    sudo systemctl enable dnsmasq
-    sudo systemctl restart dnsmasq
-    #sudo systemctl disable dnsmasq
-    #sudo systemctl stop dnsmasq
+    if [ -f /etc/rpi-issue ]; then
+      sudo apt install -y $1
+      sudo $STOW -t /etc $1
+      sudo systemctl enable $1
+      sudo systemctl restart $1
+    fi
     ;;
   "polipo")
-    sudo apt install -y polipo
-    sudo $STOW -d $PKGDIR -t /etc/polipo polipo
-    sudo systemctl enable polipo
-    sudo systemctl restart polipo
+    if [ -f /etc/rpi-issue ]; then
+      sudo apt install -y $1
+      sudo $STOW -t /etc/polipo $1
+      sudo systemctl enable $1
+      sudo systemctl restart $1
+    fi
     ;;
   "privoxy")
-    sudo apt install -y privoxy
-    [ -f /etc/privoxy/config ] && sudo rm /etc/privoxy/config
-    [ -f /etc/privoxy/user.action ] && sudo rm /etc/privoxy/user.action
-    [ -f /etc/privoxy/user.filter ] && sudo rm /etc/privoxy/user.filter
-    sudo $STOW -d $PKGDIR -t /etc/privoxy privoxy
-    sudo systemctl enable privoxy
-    sudo systemctl restart privoxy
+    if [ -f /etc/rpi-issue ]; then
+      sudo apt install -y $1
+      [ -f /etc/privoxy/config ] && sudo rm /etc/privoxy/config
+      [ -f /etc/privoxy/user.action ] && sudo rm /etc/privoxy/user.action
+      [ -f /etc/privoxy/user.filter ] && sudo rm /etc/privoxy/user.filter
+      sudo $STOW -t /etc/privoxy $1
+      sudo systemctl enable $1
+      sudo systemctl restart $1
+    fi
     ;;
   "nginx")
-    sudo apt install -y nginx
-    sudo $STOW -d $PKGDIR -t /etc/nginx nginx
-    sudo systemctl enable nginx
-    sudo systemctl restart nginx
+    if [ -f /etc/rpi-issue ]; then
+      sudo apt install -y $1
+      sudo $STOW -t /etc/nginx $1
+      sudo systemctl enable $1
+      sudo systemctl restart $1
+    fi
     ;;
   "disable_nvaudio")
-    # TODO: only for debian dev box
-    if [ ! -f /etc/modprobe.d/blacklist-nvaudio.conf ]; then
-      echo 'blacklist snd-hda-intel\n' | sudo tee /etc/modprobe.d/blacklist-nvaudio.conf
-      touch "$REBOOT_FLAG"
+    if [ -f /etc/debian_version ]; then
+      if [ ! -f /etc/modprobe.d/blacklist-nvaudio.conf ]; then
+        echo 'blacklist snd-hda-intel\n' | sudo tee /etc/modprobe.d/blacklist-nvaudio.conf
+        touch "$REBOOT_FLAG"
+      fi
     fi
     ;;
   "pulse_audio")
     # TODO: pulse audio setup for debian dev box
+    sudo apt install -y pulseaudio pavucontrol
     ;;
   "dropbox")
     # TODO: dropbox for debian dev box
     # TODO: owncloud server
     ;;
   "spotify")
-    sudo apt-key adv --keyserver hkp://keyserver.ubuntu.com:80 --recv-keys 931FF8E79F0876134EDDBDCCA87FF9DF48BF1C90
-    echo deb http://repository.spotify.com stable non-free | sudo tee /etc/apt/sources.list.d/spotify.list
-    sudo apt update
-    sudo apt install spotify-client
+    if [ -f /etc/debian_version ]; then
+      repo = 'deb http://repository.spotify.com stable non-free'
+      sudo apt-key adv --keyserver hkp://keyserver.ubuntu.com:80 \
+        --recv-keys 931FF8E79F0876134EDDBDCCA87FF9DF48BF1C90
+      echo $repo | sudo tee /etc/apt/sources.list.d/spotify.list
+      sudo apt update
+      sudo apt install spotify-client
+    fi
     ;;
   "terminator")
-    # TODO: terminatorsetup for debian dev box
+    if [ -f /etc/debian_version ]; then
+      echo 'TODO'
+    fi
     ;;
   "i3")
-    sudo apt install i3
-    [ -f $HOME/.config/i3/config ] && \
-      mv $HOME/.config/i3/config $HOME/.config/i3/config.thetapibak
-    $STOW -d $PKGDIR -t $HOME/.config/i3 $1
+    if [ -f /etc/debian_version ]; then
+      sudo apt install i3
+      [ -f $HOME/.config/i3/config ] && \
+        mv $HOME/.config/i3/config $HOME/.config/i3/config.thetapibak
+      $STOW -t $HOME/.config/i3 $1
+    fi
     ;;
   "debian_gui")
     # TODO: debian stuff for dev box
@@ -132,7 +155,7 @@ case $1 in
     # firefox-esr
     ;;
   *)
-    $STOW -d $PKGDIR -t $HOME $1
+    $STOW -t $HOME $1
     ;;
 esac
